Update Reason version & implement GADT for Lwt_unix.* / Unix.* types
In this commit we update ReasonNode's reason dependency version to
1.13.7, which is the latest release. We also implement GADT's for
Lwt_unix.* and Unix.* types in order to provide camelcased types same as
the methods from our modules. For newcomers this is also useful so they
don't have to go looking through stdlib or unix docs to find what they
are looking for.